Module reth_network::message

Capability messaging

An RLPx stream is multiplexed via the prepended message-id of a framed message. Capabilities are exchanged via the RLPx Hello message as pairs of (id, version), https://github.com/ethereum/devp2p/blob/master/rlpx.md#capability-messaging

Structs

NewBlockMessage

Internal form of a NewBlock message

PeerRequestSender

A Cloneable connection for sending requests directly to the session of a peer.

Enums

BlockRequest

Request Variants that only target block related data.

PeerMessage

All Bi-directional eth-message variants that can be sent to a session or received from a session.

PeerRequest

Protocol related request messages that expect a response

PeerResponse

Corresponding variant for PeerRequest.

PeerResponseResult

All response variants for PeerResponse