Module reth_network::peers

Peer related implementations

Structs

ConnectionInfo

Tracks stats about connected nodes

Peer

Tracks info about a single peer.

PeersConfig

Config type for initiating a PeersManager instance.

PeersHandle

A communication channel to the PeersManager to apply manual changes to the peer set.

PeersManager

Maintains the state of all the peers known to the network.

ReputationChangeWeights

How the ReputationChangeKind are weighted.

Enums

PeerAction

Actions the peer manager can trigger.

PeerKind

Represents the kind of peer