Struct reth_network::NetworkHandle

pub struct NetworkHandle { /* private fields */ }

A shareable network frontend. Used to interact with the network.

See also NetworkManager.

Implementations

impl NetworkHandle

pub fn peer_id(&self) -> &PeerId

Returns the PeerId used in the network.

pub fn peers_handle(&self) -> &PeersHandle

Returns the PeersHandle that can be cloned and shared.

The PeersHandle can be used to interact with the network’s peer set.

pub async fn fetch_client(&self) -> Result<FetchClient, RecvError>

Returns a new FetchClient that can be cloned and shared.

The FetchClient is the entrypoint for sending requests to the network.

pub fn mode(&self) -> &NetworkMode

Returns the mode of the network, either pow, or pos

pub fn update_status(&self, head: Head)

Update the status of the node.

pub fn announce_block(&self, block: NewBlock, hash: B256)

Announce a block over devp2p

Caution: in PoS this is a noop, since new block are no longer announced over devp2p, but are instead sent to node node by the CL. However, they can still be requested over devp2p, but broadcasting them is a considered a protocol violation..

pub fn send_request(&self, peer_id: PeerId, request: PeerRequest)

Sends a PeerRequest to the given peer’s session.

pub fn send_transactions_hashes( &self, peer_id: PeerId, msg: NewPooledTransactionHashes )

Send transactions hashes to the peer.

pub fn send_transactions( &self, peer_id: PeerId, msg: Vec<Arc<TransactionSigned>> )

Send full transactions to the peer

pub fn bandwidth_meter(&self) -> &BandwidthMeter

Provides a shareable reference to the BandwidthMeter stored on the NetworkInner.

pub async fn shutdown(&self) -> Result<(), RecvError>

Send message to gracefully shutdown node.

This will disconnect all active and pending sessions and prevent new connections to be established.

pub fn set_network_active(&self)

Set network connection state to Active.

New outbound connections will be established if there’s capacity.

pub fn set_network_hibernate(&self)

Set network connection state to Hibernate.

No new outbound connections will be established.

pub fn tx_gossip_disabled(&self) -> bool

Whether tx gossip is disabled

pub fn secret_key(&self) -> &SecretKey

Returns the secret key used for authenticating sessions.